leave_automation v1.1/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
gspread==5.7.0
oauth2client==4.1.3
openai==2.8.1   # or other official client; adjust version if needed